requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==3.0.3
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
msgpack==0.6.2
msgpack-python==0.5.6
Pillow==7.0.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2019.3
redis==2.10.6
requests==2.23.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
stripe==2.43.0
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.8
zope.interface==4.7.2
gunicorn==20.0.4
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
librosa==0.6.3
numba==0.47.0
numpy==1.16.0
pandas==1.0.3
pytest==5.4.1
tensorflow==1.15.2
djangorestframework==3.11.0